+++ src/complex/arith.rs
@@ -824,6 +824,8 @@ mod tests {
     #[test]
     fn check_arith_others() {
         use tests::{F32, F64, I32, I64, U32, U64};
+        #[cfg(int_128)]
+        use tests::{I128, U128};
         let large = [
             Complex::with_val(20, (Special::Zero, 1.0)),
             Complex::with_val(20, (Special::NegZero, 1.0)),
@@ -872,6 +874,11 @@ mod tests {
             .chain(F32.iter().map(|&x| Complex::with_val(20, x)))
             .chain(F64.iter().map(|&x| Complex::with_val(20, x)))
             .collect::<Vec<Complex>>();
+        #[cfg(int_128)]
+        {
+            against.extend(U128.iter().map(|&x| Complex::with_val(20, x)));
+            against.extend(I128.iter().map(|&x| Complex::with_val(20, x)));
+        }
         #[cfg(feature = "integer")]
         against.extend(z.iter().map(|x| Complex::with_val(20, x)));
         #[cfg(feature = "rational")]