+++ src/float/arith.rs
@@ -799,6 +799,8 @@ pub(crate) mod tests {
     #[test]
     fn check_arith_others() {
         use tests::{F32, F64, I32, I64, U32, U64};
+        #[cfg(int_128)]
+        use tests::{I128, U128};
         let large = [
             Float::with_val(20, Special::Zero),
             Float::with_val(20, Special::NegZero),
@@ -839,12 +841,17 @@ pub(crate) mod tests {
             .chain(F32.iter().map(|&x| Float::with_val(20, x)))
             .chain(F64.iter().map(|&x| Float::with_val(20, x)))
             .collect::<Vec<Float>>();
-        #[cfg(feature = "integer")]
+        #[cfg(any(int_128, feature = "integer"))]
         let mut against = against;
         #[cfg(feature = "integer")]
         against.extend(z.iter().map(|x| Float::with_val(20, x)));
         #[cfg(feature = "rational")]
         against.extend(q.iter().map(|x| Float::with_val(20, x)));
+        #[cfg(int_128)]
+        {
+            against.extend(U128.iter().map(|&x| Float::with_val(20, x)));
+            against.extend(I128.iter().map(|&x| Float::with_val(20, x)));
+        }
 
         for op in U32 {
             let fop = Float::with_val(100, *op);