+++ src/float/cmp.rs
@@ -145,6 +145,8 @@ cmp_i! { i64, |f, t: &i64| unsafe { xmpfr::cmp_i64(f, *t) } }
 cmp_i! { isize, |f, t: &isize| unsafe { mpfr::cmp_si(f, cast(*t)) } }
 #[cfg(target_pointer_width = "64")]
 cmp_i! { isize, |f, t: &isize| unsafe { xmpfr::cmp_i64(f, cast(*t)) } }
+#[cfg(int_128)]
+cmp_i! { i128, |f, t: &i128| unsafe { xmpfr::cmp_i128(f, *t) } }
 
 cmp_i! { u8, |f, t: &u8| unsafe { mpfr::cmp_ui(f, cast(*t)) } }
 cmp_i! { u16, |f, t: &u16| unsafe { mpfr::cmp_ui(f, cast(*t)) } }
@@ -154,6 +156,8 @@ cmp_i! { u64, |f, t: &u64| unsafe { xmpfr::cmp_u64(f, *t) } }
 cmp_i! { usize, |f, t: &usize| unsafe { mpfr::cmp_ui(f, cast(*t)) } }
 #[cfg(target_pointer_width = "64")]
 cmp_i! { usize, |f, t: &usize| unsafe { xmpfr::cmp_u64(f, cast(*t)) } }
+#[cfg(int_128)]
+cmp_i! { u128, |f, t: &u128| unsafe { xmpfr::cmp_u128(f, *t) } }
 
 cmp_f! { f32, |f, t: &f32| unsafe { mpfr::cmp_d(f, cast(*t)) } }
 cmp_f! { f64, |f, t: &f64| unsafe { mpfr::cmp_d(f, *t) } }
@@ -205,7 +209,7 @@ mod tests {
         T: Copy + PartialEq<Float> + PartialOrd<Float>,
     {
         for op in s {
-            let fop = Float::with_val(100, *op);
+            let fop = Float::with_val(150, *op);
             for b in against {
                 assert_eq!(b.eq(op), <Float as PartialEq>::eq(&b, &fop));
                 assert_eq!(op.eq(&b), <Float as PartialEq>::eq(&fop, &b));
@@ -233,7 +237,7 @@ mod tests {
         T: PartialEq<Float> + PartialOrd<Float>,
     {
         for op in s {
-            let fop = Float::with_val(100, op);
+            let fop = Float::with_val(150, op);
             for b in against {
                 assert_eq!(b.eq(op), <Float as PartialEq>::eq(&b, &fop));
                 assert_eq!(op.eq(&b), <Float as PartialEq>::eq(&fop, &b));
@@ -257,6 +261,8 @@ mod tests {
     #[test]
     fn check_cmp_others() {
         use tests::{F32, F64, I32, I64, U32, U64};
+        #[cfg(int_128)]
+        use tests::{I128, U128};
         let large = [
             Float::with_val(20, Special::Zero),
             Float::with_val(20, Special::NegZero),
@@ -297,12 +303,19 @@ mod tests {
             .chain(F32.iter().map(|&x| Float::with_val(20, x)))
             .chain(F64.iter().map(|&x| Float::with_val(20, x)))
             .collect::<Vec<Float>>();
-        #[cfg(feature = "integer")]
+        #[cfg(any(int_128, feature = "integer"))]
         let mut against = against;
         #[cfg(feature = "integer")]
         against.extend(z.iter().map(|x| Float::with_val(20, x)));
         #[cfg(feature = "rational")]
         against.extend(q.iter().map(|x| Float::with_val(20, x)));
+        #[cfg(int_128)]
+        {
+            against.extend(U128.iter().map(|&x| Float::with_val(20, x)));
+            against.extend(I128.iter().map(|&x| Float::with_val(20, x)));
+            check_cmp_prim(U128, &against);
+            check_cmp_prim(I128, &against);
+        }
         check_cmp_prim(U32, &against);
         check_cmp_prim(I32, &against);
         check_cmp_prim(U64, &against);