+++ src/misc.rs
@@ -50,7 +50,11 @@ macro_rules! neg_abs_signed {
     )* };
 }
 neg_abs_unsigned! { u8 u16 u32 u64 usize }
+#[cfg(int_128)]
+neg_abs_unsigned! { u128 }
 neg_abs_signed! { (i8, u8) (i16, u16) (i32, u32) (i64, u64) (isize, usize) }
+#[cfg(int_128)]
+neg_abs_signed! { (i128, u128) }
 
 #[allow(unknown_lints, transmute_int_to_float)]
 pub fn trunc_f64_to_f32(f: f64) -> f32 {