+++ RELEASES.md
@@ -1,3 +1,4 @@
+* Add `i128` and `u128` support, conditional on compiler support.
 * Add `Float::get_significand` method.
 
 Version 1.0.1 (2018-03-10)